The Windows build named the files that existed when it was written
`nmake /f Makefile.msvc' could not have linked, and had not been able to for
as long as the file has existed.  Its object list predates src/boot, so
nothing in it compiled bootstrap.c or profile.c -- and src/sched/worker.c,
src/gfx/font_face.c and four of the seven files under src/compiler were
missing too.  Eight translation units.  main.obj referenced WORKER_self,
st_safepoint_requested, BOOT_bootstrap and the rest, and the link ended in
unresolved externals.

Checked rather than assumed: compiling exactly the old set plus src/main.c
with gcc gives 40 undefined references.  The corrected set links clean under
both OM=mt and OM=bb.  nmake has no wildcard, so the GNU makefile picks up a
new .c under src/ and this one does not; the list now says so and says to
add to both.

MSVC KEEPS ITS C11 ATOMICS BEHIND A FLAG.  src/port/st_atomic.h includes
<stdatomic.h> directly, and cl answers

    vcruntime_c11_stdatomic.h(12): fatal error C1189: #error:
        "C atomic support is not enabled"

which is not what "MSVC has supported C11 atomics since 17.5" leads anyone
to expect.  /experimental:c11atomics is still required -- confirmed against
MSVC 14.50 under Visual Studio 18, which is to say it has not stopped being
required in the three toolsets since.  Set for cl and not for clang-cl,
which brings the real gcc/clang atomics and rejects the flag.

C4333 ON EVERY TRANSLATION UNIT.  ST_header_frame_size shifts an st_oop
right by 16, and under OM=bb an st_oop is sixteen bits.  The answer there is
always zero, which is the design -- "not stated", fall back to the Blue
Book's small/large bit -- and integer promotion to int makes it defined
rather than undefined.  cl is right about the data and wrong about the loss.
Widened before the shift so the intent is stated in the code; both compilers
fold it to 0 under bb, and nobody has to reach for /wd4333, which would also
hide the next shift that really is a mistake.

Four more, none of them the reason it failed: OM now defaults to mt as the
GNU makefile has since `make gives you the system', and the Blue Book memory
cannot bootstrap an image -- which is the same two-layers-below failure that
commit removed on the other side.  SDL3's import library defaults to
lib\x64\SDL3.lib, where libsdl.org's -VC.zip puts it, with SDL3LIB= for
vcpkg's flat lib\.  /MD, because SDL3 is a DLL against the dynamic CRT and
cl defaults to the static one.  And `test' builds all sixteen suites instead
of a chosen two -- each is wrapped in #ifdef ST_OM_MT or ST_OM_BB and
reports itself skipped under the other memory, so one list is right for both
builds.  Written out a line per suite: nmake and cmd disagree about how many
percent signs a `for' variable wants, and a test runner is the last place to
find out which of them was right.

sched_yield had no header.  src/port/st_port.c calls it and reached the
declaration through glibc's <pthread.h>, which is a property of glibc and
not a promise POSIX makes.  With -Werror=implicit-function-declaration in
the warning set, the first C library that does not do the same favour is a
build that stops rather than a warning.  Included explicitly.

Windows.md and macOS.md are the builds written down, each with a section
separating what was checked from what was not.  Neither platform has
compiled this and both say so twice.  What was checked for macOS, on Linux:
`make UNAME_S=Darwin -Bn' emits zero occurrences of pthread where the Linux
run emits 23; everything builds and links under -U__linux__, which takes the
same three branches macOS takes, with both memories and all sixteen suites
and the benchmark clean under the production warning set; the makefile uses
no GNU make 4.x construct, so Apple's 3.81 is enough; tests/run_profiles.sh
is POSIX sh with no GNU-isms, so unlike Windows the whole of `make test'
runs natively.  Two things macOS reports rather than fixes: `make bench'
reads low on Apple Silicon, because pin_to_physical_cores is __linux__-only
and Darwin has neither /sys nor sched_setaffinity, and the window is created
without SDL_WINDOW_HIGH_PIXEL_DENSITY, so Retina scaling is the window
server's.  Windows reports one: src/boot/profile.c splits paths on '/' only,
so -profile profiles\st2026.profile resolves #requires against the wrong
directory and forward slashes are what to type.

.gitignore gains st80.exe, *.obj, *.pdb and *.ilk: `st80' does not cover the
first and `*.o' does not cover the second.

All sixteen unit suites pass under both object memories after the two source
changes.

+++ src/interp/interp.h
@@ -198,7 +198,25 @@ ST_header_large_context(st_oop header)
 static inline unsigned
 ST_header_frame_size(st_oop header)
 {
-    return (unsigned) ((header >> ST_HEADER_FRAME_SHIFT) & 0xFF);
+    /*
+     *  Widened before the shift, and the cast is not decoration.
+     *
+     *  Under OM=bb an st_oop is sixteen bits and this field starts at bit
+     *  16, so the answer there is always zero -- which is the design stated
+     *  above ("not stated", fall back to the Blue Book's small/large bit)
+     *  and not an accident.  Shifting the narrow type says exactly that to
+     *  the machine, because the integer promotion to int makes it defined,
+     *  and something alarming to the reader: MSVC answers C4333, "right
+     *  shift by too large amount, data loss", once for every translation
+     *  unit that includes this header.  It is right about the data and
+     *  wrong about the loss.
+     *
+     *  Saying the widening out loud costs nothing -- both compilers fold
+     *  the whole expression to 0 under bb -- and it keeps the warning off
+     *  without anyone reaching for /wd4333, which would also hide the next
+     *  shift that really is a mistake.
+     */
+    return (unsigned) (((uint64_t) header >> ST_HEADER_FRAME_SHIFT) & 0xFF);
 }
 
 /*
